refactor: 관리자와 사용자 갤러리 공통 기능을 갤러리 클래스로 생성 추후 확장성(관리자 페이지의 카드 추가, 수정, 삭제 기능)을 고려하여 분리시킴 GitHub issue #210
