+++ src/pages/gallery/Gallery.js
@@ -1,0 +1,57 @@
+import axios from 'axios';
+import './gallery.css';
+
+// AdminGallery.js와 Employee.js의 공통 기능을 추출해 기본 클래스로 생성
+export class Gallery {
+  constructor(container, props = {}) {
+    this.container.container;
+    this.apiEndpoint = props.galleryDataPath || '/api/v1/gallery';
+  }
+
+  async getGalleryData() {
+    try {
+      const response = await axios.get(this.apiEndpoint);
+      return response.data.data;
+    } catch (e) {
+      console.error('gallery contents를 fetch해서 불러오는 데 실패했습니다.', e);
+      return [];
+    }
+  }
+  getImageUrl(url) {
+    return new URL(`${url}`, import.meta.url).href;
+  }
+
+  async updateGalleryContainer() {
+    const galleryCardData = await this.getGalleryData();
+    this.renderGalleryContainer(galleryCardData);
+  }
+
+  render() {
+    this.container.innerHTML = /* HTML */ `
+      <div class="gallery">
+        <h1 class="gallery__title">갤러리 관리</h1>
+        <div class="gallery__container"></div>
+      </div>
+    `;
+    this.updateGalleryContainer();
+  }
+
+  renderGalleryContainer(galleryCardData) {
+    const galleryContainer = this.container.querySelector('.gallery__container');
+    galleryContainer.innerHTML = /* HTML */ `
+      ${galleryCardData.map((item) => this.cardTemplate(item)).join('')}
+    `;
+  }
+
+  cardTemplate({ image, title, date }) {
+    return /* HTML */ `
+      <div class="gallery__container-card">
+        <div class="gallery__container-image-area">
+          <img src="${this.getImageUrl(image) || 'no images'}" alt="${title}" />
+        </div>
+        <div class="gallery__container-title">${title || '제목 없음'}</div>
+        <div class="gallery__container-date">${date}</div>
+      </div>
+    `;
+  }
+}
