+++ src/pages/index.js
@@ -1,4 +1,5 @@
-// import 순서를 관리자와 사용자별 메뉴 순서에 맞게 수정했습니다.
+// 로딩 컴포넌트 추가
+import Loading from '../components/loading/Loading.js';
 import { Layout } from './layout/Layout.js';
 import { EmployeeListTable } from './employeeListTable/EmployeeListTable.js';
 // import { AttendanceList } from './attendanceList/AttendanceList.js';
@@ -15,6 +16,7 @@ import Login from './login/userLogin.js';
 import LeaveApplicationList from './LeaveApplicationList/LeaveApplicationList.js';
 
 const app = document.querySelector('#app');
+const loading = new Loading(app);
 
 // isAdmin으로 admin 상태를 확인
 const isAdmin = sessionStorage.getItem('admin') === 'true';
@@ -90,25 +92,36 @@ if (!sessionStorage.id) {
     },
   };
 
-  const renderComponent = ({ ComponentClass, props }) => {
+  const renderComponent = async ({ ComponentClass, props }) => {
     const componentInstance = new ComponentClass(routeView, props);
     componentInstance.render();
+
+    if (typeof componentInstance.dataLoaded === 'function') {
+      await componentInstance.dataLoaded();
+    }
   };
 
-  function router(props = {}) {
+  async function router(props = {}) {
     const routes = isAdmin ? adminRoutes : userRoutes;
     const view = routes[location.pathname];
     if (view) {
       const { Component: ComponentClass, title: title } = view;
       document.title = title;
       routeView.innerHTML = '';
-      renderComponent({ ComponentClass, props });
+      // 로딩 컴포넌트 표시
+      loading.render();
+
+      try {
+        await renderComponent({ ComponentClass, props });
+      } finally {
+        // 로딩 컴포넌트 숨기기
+        loading.fadeOutLoadingPage();
+      }
     } else {
       history.replaceState('', '', '/');
-      router();
+      await router();
     }
   }
-
   // Handle navigation
   window.addEventListener('click', (e) => {
     const anchorElem = e.target.closest('a');
